// backend/src/types.ts
export interface ApcUpsStatusProps {
    APC: string;
    DATE: string;
    HOSTNAME: string;
    VERSION: string;
    UPSNAME: string;
    CABLE: string;
    DRIVER: string;
    UPSMODE: string;
    STARTTIME: string;
    MODEL: string;
    STATUS: string;
    LINEV: string;
    LOADPCT: string;
    BCHARGE: string;
    TIMELEFT: string;
    MBATTCHG: string;
    MINTIMEL: string;
    MAXTIME: string;
    SENSE: string;
    LOTRANS: string;
    HITRANS: string;
    ALARMDEL: string;
    BATTV: string;
    LASTXFER: string;
    NUMXFERS: string;
    TONBATT: string;
    CUMONBATT: string;
    XOFFBATT: string;
    SELFTEST: string;
    STATFLAG: string;
    SERIALNO: string;
    BATTDATE: string;
    NOMINV: string;
    NOMBATTV: string;
    NOMPOWER: string;
    FIRMWARE: string;
    "END APC": string;
}